+++ frontend/package.json
@@ -19,6 +19,7 @@
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.0.0-beta.0",
     "react-scripts": "4.0.3",
+    "react-toastify": "^7.0.4",
     "typescript": "^4.1.2",
     "web-vitals": "^1.0.1"
   },
