+++ frontend/src/reducer/Reducer.ts
@@ -1,12 +1,11 @@
 import { GET_ALL_CATEGORIES } from '../constants/constants';
-import { StateInterface } from '../types/context.types';
-import { ActionsTypes } from '../types/reducer.types';
+import { StateInterface, ActionsTypes } from '../types/types';
 
 export const reducer = (state: StateInterface, action: ActionsTypes) => {
     switch (action.type) {
         case GET_ALL_CATEGORIES:
-            console.log(action.payload, "pay");
             return state;
+        // return { ...state, categories: action.payload.categories };
         default: return state;
     }
 };
