--- frontend/src/types/context.types.ts
@@ -1,12 +1,0 @@
-import React from "react";
-import { ActionsTypes } from "./reducer.types";
-
-export interface StateInterface {
-    user: {};
-    categories: [];
-}
-
-export interface ContextValue {
-    state: StateInterface;
-    dispatch: (action: ActionsTypes) => void;
-}
