+++ frontend/src/types/types.ts
@@ -1,0 +1,38 @@
+import { GET_ALL_CATEGORIES } from '../constants/constants';
+
+// ----------------------------------------------------CONTEXT--------------------------------------------------
+
+export interface StateInterface {
+    user: {};
+    categories: [];
+}
+
+export interface ContextValue {
+    state: StateInterface;
+    dispatch: (action: ActionsTypes) => void;
+}
+
+// -----------------------------------------------------REDUCER--------------------------------------------------
+
+export type ActionsTypes =
+    | {
+        type: typeof GET_ALL_CATEGORIES,
+        payload: { categories: Category[]; };
+    };
+
+//---------------------------------------------------------CATEGORY---------------------------------------------
+
+export interface Category {
+    level: string;
+    image: string;
+    quizzes: string[];
+}
+
+export interface ServerError {
+    success: boolean;
+    message: string;
+}
+export interface CategoriesRes {
+    success: boolean;
+    categories: Category[];
+}
